requirements.txt
fastapi==0.106.0
fastapi_pagination==0.12.14
sqlalchemy==2.0.23
asyncpg==0.29.0
pytest==7.4.3
pytest-asyncio==0.23.2
httpx==0.26.0
rich==13.7.0
pydantic-settings==2.1.0
alembic==1.13.1
black==23.12.1
pandas==2.1.4
psycopg2==2.9.9
flake8==3.9.0  # https://github.com/flakehell/flakehell/issues/10
flake8-black==0.3.2
flake8-bugbear==22.4.25
flake8-comprehensions==3.8.0
flake8-docstrings==1.6.0
flake8-isort==4.1.1
flake8-pytest-style==1.6.0
flakehell==0.9.0
jinja2==3.1.2
mypy==1.8.0